Setting is_component_build to false by default
Webrtc does not support component builds so we want to override
the chromium default value (which can be true on debug builds if the
os is different from iOS).

Please note that the user can set this value to true in two ways:

 - using --args (e.g.: gn gen out/default --args='is_component_build=true'
 - changing the value in the args.gn file

But in both cases the value will be ignored because we don't use the
'component' template but we rely directly on 'rtc_static_library' and
'rtc_shared_library'.

BUG=webrtc:6975
NOTRY=True

Review-Url: https://codereview.webrtc.org/2728643003
Cr-Original-Commit-Position: refs/heads/master@{#17020}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 2cb3944ba7e4467601ebd94888e1811a6d77be49