Add |protected_by_flexfec| flag to VideoReceiveStream::Config.
Use of FlexFEC is known when streams are created in
WebRtcVideoChannel2, so this replaces the code in Call to infer
FlexFEC config of video streams from the configuration of the FlexFEC
stream(s). This also allows us to switch to a more logical creation
order, where media streams are created before the FlexFEC stream.

This is done in preparation for a larger refactoring of the RTP
demuxing done in Call.

BUG=None

Review-Url: https://codereview.webrtc.org/2712683002
Cr-Original-Commit-Position: refs/heads/master@{#17143}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: c69385de8b614b7edba4f640bca6aa0963e1505c