Removing test deps in webrtc/test/fuzzers
Targets in webrtc/test/fuzzers are used in chromium which includes WebRTC
with rtc_include_tests=false.

We enabled 'gn check' on the webrtc/test directory and we have detected
that some dependencies were not tracked. These dependencies are on test
targets so we cannot add them in the dep list because this causes a
breakage in chromium.

BUG=webrtc:7515
NOTRY=True

Review-Url: https://codereview.webrtc.org/2840523003
Cr-Original-Commit-Position: refs/heads/master@{#17844}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 14b86d386460a759da1d5029cc09eda46330ea2b